requirements.txt
flask==3.0.0
gunicorn==21.2.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
tensorflow-cpu>=2.15.0
Werkzeug==3.0.1
